pyproject.toml
[project]
name = "iqtidor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-unfold>=0.62.0",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg>=3.2.9",
    "python-decouple>=3.8",
]
